requirements.txt
web3
litellm
python-dotenv